Added unit test all cases check accumulated between 0 and 100000

+++ src/api/Services/TransferService.cs
@@ -5,7 +5,7 @@ namespace api.Services
 {
     public class TransferService
     {
-        public bool IsAccumulated(int accumulatedNumber)
+        public bool IsAccumulated(double accumulatedNumber)
         {
             if(accumulatedNumber == 0) {
                 return true;
